+++ readium/streamer/pdf/pdfium/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="org.readium.r2.streamer.pdf.pdfium" />
