+++ readium/streamer/pdf/src/main/java/org/readium/r2/streamer/pdf/PdfPositionsService.kt
@@ -1,19 +1,10 @@
-/*
- * Module: r2-streamer-kotlin
- * Developers: Mickaël Menu
- *
- * Copyright (c) 2020. Readium Foundation. All rights reserved.
- * Use of this source code is governed by a BSD-style license which is detailed in the
- * LICENSE file present in the project repository where this source code is maintained.
- */
+package org.readium.r2.streamer.pdf
 
-package org.readium.r2.streamer.parser.pdf
-
-import org.readium.r2.shared.util.mediatype.MediaType
 import org.readium.r2.shared.publication.Link
 import org.readium.r2.shared.publication.Locator
 import org.readium.r2.shared.publication.Publication
 import org.readium.r2.shared.publication.services.PositionsService
+import org.readium.r2.shared.util.mediatype.MediaType
 import timber.log.Timber
 
 /**
